feat: change settings functionality to separate file
