+++ index.html
@@ -183,5 +183,6 @@
             </div>
         </div>
         <script type="module" src="src/main.js"></script>
+        <script type="module" src="src/utils/settings.js"></script>
     </body>
 </html>
